Allow PreferenceScreens to have a key
Being able to tag PreferenceScreens with a key is useful for organizational reasons, even if they don't persist any data to the SharedPreferences.
For backwards compatibility and flexibility, the default key for PreferenceScreens is still an empty string.
The root screen, however, now gets a special reserved key, which may not be used for sub-screens (but could theoretically still be used for Preferences, this is highly discouraged though). With that root key, it will be possible to do better screen traversal in the future.