Fix not being able to set attributes on collapse blocks
Problem: the collapse DSL function had to use "PreferenceScreen.Builder" as a λ-function receiver to be able to add further preferences into the screen - however, this meant any setters meant for the CollapsePreference were executed on the PreferenceScreen.

Solution: introduce a PreferenceScreen.Appendable interface, which receives the calls to add further preferences, and proxy it to the actual preference screen in CollapsePreference. This has the additional benefit that the collapse preference now directly gets the list of its children, without requiring an extra callback from the preference screen.