Reference preferences in DependencyManager weakly to prevent leaks
Preferences themselves are safe against Context leaks, however, if the preference tree is always recreated in line with the Activity and the PreferenceAdapter holds a reference to the Activity (in form of a listener), hard-referencing the preferences here can leak the Activity.
Additionally, always keeping all Preferences in memory might be undesirable.